Adding create-user command
